[Silver III] Title: 주유소, Time: 380 ms, Memory: 36588 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -6,10 +6,10 @@ public class Main {
         int N = Integer.parseInt(br.readLine());
         StringTokenizer st;
 
-        int[] oil = new int[N-1];
-        int[] price = new int[N];
+        long[] oil = new long[N-1];
+        long[] price = new long[N];
 
-        int totalOil=0;
+        long totalOil=0;
         st = new StringTokenizer(br.readLine()," ");
         for(int i=0; i<N-1; i++){
             oil[i]=Integer.parseInt(st.nextToken());
@@ -19,7 +19,7 @@ public class Main {
             price[i]=Integer.parseInt(st.nextToken());
         }
 
-        int maxOil = price[0];
+        long maxOil = price[0];
         for(int i=0; i<N-1; i++){
             if(maxOil>price[i]){
                 maxOil=price[i];
